Fixed warning by adding an explicit cast to long; TimerThread.c ..\..\threadutil\src\TimerThread.c(110): warning C4244: '=' : conversion from 'time_t' to 'long', possible loss of data

+++ threadutil/src/TimerThread.c
@@ -107,7 +107,7 @@ static void *TimerThreadWorker(
         }
         if (nextEvent) {
             timeToWait.tv_nsec = 0;
-            timeToWait.tv_sec = nextEvent->eventTime;
+            timeToWait.tv_sec = (long)nextEvent->eventTime;
             ithread_cond_timedwait( &timer->condition, &timer->mutex,
                                     &timeToWait );
         } else {
